+++ source/databricks/calculation-engine/package/steps/aggregation/transformations/__init__.py
@@ -17,7 +17,7 @@ from .aggregation_result_formatter import (
 )
 from .aggregate_quality import aggregate_total_consumption_quality, aggregate_quality
 from .aggregate_sum_and_set_quality import aggregate_sum_and_set_quality
-from .adjust_grid_loss import adjust_production, adjust_flex_consumption
+from .apply_grid_loss_adjustment import adjust_production, adjust_flex_consumption
 from .combine_master_data import (
     combine_negative_grid_loss_with_master_data,
     combine_positive_grid_loss_with_master_data,
